+++ python/tests/test_upsample_2d_fourier.py
@@ -35,8 +35,8 @@ class Upsample2DFourierTests(unittest.TestCase):
             torch.ops.load_library(str(args.kernel_lib))
         cls.compiled_f32 = torch._inductor.aoti_load_package(str(args.package_f32))
         cls.compiled_f64 = torch._inductor.aoti_load_package(str(args.package_f64))
-        cls.eager_f32 = Upsample2DFourier(input_dtype=torch.float32).eval()
-        cls.eager_f64 = Upsample2DFourier(input_dtype=torch.float64).eval()
+        cls.eager_f32 = Upsample2DFourier().eval()
+        cls.eager_f64 = Upsample2DFourier().eval()
 
     def _make_inputs(self):
         py_input = torch.tensor(
